+++ cross-compile/onvif/src/protocol/gsoap/onvif_gsoap_core.c
@@ -137,7 +137,7 @@ int onvif_gsoap_init_request_parsing(onvif_gsoap_context_t* ctx, const char* req
 
   /* Validate request size to prevent malloc corruption */
   if (xml_size > MAX_ONVIF_REQUEST_SIZE) {
-    platform_log_debug("onvif_gsoap_init_request_parsing: Request too large - xml_size=%zu, max=%d", xml_size, MAX_ONVIF_REQUEST_SIZE);
+    platform_log_debug("onvif_gsoap_init_request_parsing: Request too large - xml_size=%zu, max=%zu", xml_size, (size_t)MAX_ONVIF_REQUEST_SIZE);
     onvif_gsoap_set_error(ctx, ONVIF_ERROR_INVALID, __func__, "Request size exceeds maximum allowed size (1MB)");
     return ONVIF_ERROR_INVALID;
   }
